+++ Leavepal/src/main/java/org/kumaran/entity/LeaveTrackerData.java
@@ -30,27 +30,27 @@ public class LeaveTrackerData {
 
     @Column(nullable = false)
     @Schema(description = "Sick leave available in current cycle", example = "4")
-    private Integer sickLeaveAvailable = 0;
+    private Double sickLeaveAvailable = 0.0;
 
     @Column(nullable = false)
     @Schema(description = "Casual leave available in current cycle", example = "4")
-    private Integer casualLeaveAvailable = 0;
+    private Double casualLeaveAvailable = 0.0;
 
     @Column(nullable = false)
     @Schema(description = "Loss of pay available", example = "0")
-    private Integer lossOfPayAvailable = 0;
+    private Double lossOfPayAvailable = 0.0;
 
     @Column(nullable = false)
     @Schema(description = "Sick leave booked", example = "0")
-    private Integer sickLeaveBooked = 0;
+    private Double sickLeaveBooked = 0.0;
 
     @Column(nullable = false)
     @Schema(description = "Casual leave booked", example = "0")
-    private Integer casualLeaveBooked = 0;
+    private Double casualLeaveBooked = 0.0;
 
     @Column(nullable = false)
     @Schema(description = "Loss of pay booked", example = "0")
-    private Integer lossOfPayBooked = 0;
+    private Double lossOfPayBooked = 0.0;
 
     @Column(nullable = false, updatable = false)
     @Schema(description = "Created timestamp")
@@ -121,56 +121,56 @@ public class LeaveTrackerData {
         this.joiningDate = joiningDate;
     }
 
-    public Integer getSickLeaveAvailable() {
+    public Double getSickLeaveAvailable() {
         return sickLeaveAvailable;
     }
 
-    public void setSickLeaveAvailable(Integer sickLeaveAvailable) {
+    public void setSickLeaveAvailable(Double sickLeaveAvailable) {
         this.sickLeaveAvailable = sickLeaveAvailable;
         this.updatedAt = System.currentTimeMillis();
     }
 
-    public Integer getCasualLeaveAvailable() {
+    public Double getCasualLeaveAvailable() {
         return casualLeaveAvailable;
     }
 
-    public void setCasualLeaveAvailable(Integer casualLeaveAvailable) {
+    public void setCasualLeaveAvailable(Double casualLeaveAvailable) {
         this.casualLeaveAvailable = casualLeaveAvailable;
         this.updatedAt = System.currentTimeMillis();
     }
 
-    public Integer getLossOfPayAvailable() {
+    public Double getLossOfPayAvailable() {
         return lossOfPayAvailable;
     }
 
-    public void setLossOfPayAvailable(Integer lossOfPayAvailable) {
+    public void setLossOfPayAvailable(Double lossOfPayAvailable) {
         this.lossOfPayAvailable = lossOfPayAvailable;
         this.updatedAt = System.currentTimeMillis();
     }
 
-    public Integer getSickLeaveBooked() {
+    public Double getSickLeaveBooked() {
         return sickLeaveBooked;
     }
 
-    public void setSickLeaveBooked(Integer sickLeaveBooked) {
+    public void setSickLeaveBooked(Double sickLeaveBooked) {
         this.sickLeaveBooked = sickLeaveBooked;
         this.updatedAt = System.currentTimeMillis();
     }
 
-    public Integer getCasualLeaveBooked() {
+    public Double getCasualLeaveBooked() {
         return casualLeaveBooked;
     }
 
-    public void setCasualLeaveBooked(Integer casualLeaveBooked) {
+    public void setCasualLeaveBooked(Double casualLeaveBooked) {
         this.casualLeaveBooked = casualLeaveBooked;
         this.updatedAt = System.currentTimeMillis();
     }
 
-    public Integer getLossOfPayBooked() {
+    public Double getLossOfPayBooked() {
         return lossOfPayBooked;
     }
 
-    public void setLossOfPayBooked(Integer lossOfPayBooked) {
+    public void setLossOfPayBooked(Double lossOfPayBooked) {
         this.lossOfPayBooked = lossOfPayBooked;
         this.updatedAt = System.currentTimeMillis();
     }
